Add series data helper class
